requirements.txt
websocket_client
TA-Lib
numpy
